* Fixed leaks, both for greedy and unicost search types
* Compilation is run with all possible error flags

+++ SolverHeuristics.cpp
@@ -9,25 +9,20 @@ float		NPuzzle::Solver::calcHeuristic(const size_t* curr_state) const
 	switch(func_num)
 	{
 		case MANHATTAN:
-			// std::cout << "MANHATTAN" << std::endl;
 			res = NPuzzle::Solver::ManhattanDist(GetFinalState().first, curr_state, false);
 			break ;
 		case EUCLIDE:
-			// std::cout << "EUCLIDE" << std::endl;
 			res = NPuzzle::Solver::EuclideanDist(GetFinalState().first, curr_state);
 			break ;
 		case MIS_TILES:
-			// std::cout << "MIS_TILES" << std::endl;
 			res = NPuzzle::Solver::MisplacedTiles(GetFinalState().first, curr_state);
 			break ;
 		case LIN_CONF:
-			// std::cout << "LIN_CONF" << std::endl;
 			res = NPuzzle::Solver::ManhattanDist(GetFinalState().first, curr_state, true);
 			break ;
 		default:
 			throw std::string("Wrong defined heuristic function");
 	}
-	// std::cout << "res: " << res << std::endl;
 	return (res);
 }
 
@@ -48,12 +43,12 @@ size_t		ft_calc_horiz_conflicts(const size_t* curr_state, const size_t puzzle_le
 	size_t		conflicts;
 
 	conflicts = 0;
-	for (int i = 0; i < puzzle_len; ++i)
+	for (size_t i = 0; i < puzzle_len; ++i)
 	{
 		size_t		temp;
 
 		temp = 0;
-		for (int j = 0; j < puzzle_len; ++j)
+		for (size_t j = 0; j < puzzle_len; ++j)
 		{
 			const size_t		position = i * puzzle_len + j;
 			const size_t		num = curr_state[position];
@@ -75,12 +70,12 @@ size_t		ft_calc_vertic_conflicts(const size_t* curr_state, const size_t puzzle_l
 	size_t		conflicts;
 
 	conflicts = 0;
-	for (int i = 0; i < puzzle_len; ++i)
+	for (size_t i = 0; i < puzzle_len; ++i)
 	{
 		size_t		temp;
 
 		temp = 0;
-		for (int j = 0; j < puzzle_len; ++j)
+		for (size_t j = 0; j < puzzle_len; ++j)
 		{
 			const size_t		position = j * puzzle_len + i;
 			const size_t		num = curr_state[position];
@@ -115,11 +110,8 @@ float		NPuzzle::Solver::ManhattanDist(const size_t* final_state, const size_t* c
 			goal = findArrElem(final_state, GetPuzzleSize(), curr_state[i]);
 			dist += std::abs(static_cast<int>((i % GetPuzzleSize()) - goal.first))
 						+ std::abs(static_cast<int>((i / GetPuzzleSize()) - goal.second));
-			// std::cout << "MANHATTAN distance:" << dist << std::endl;		
 		}
 	}
-	// std::cout << "MANHATTAN result:" << dist << std::endl;
-	// std::cout << "LinearConflict:" << LinearConflict << std::endl;
 	if (LinearConflict)
 		dist += static_cast<int>(ft_add_lin_conf(curr_state, GetPuzzleSize()));
 	return (dist);
@@ -159,11 +151,9 @@ float		NPuzzle::Solver::EuclideanDist(const size_t* final_state, const size_t* c
 			goal = findArrElem(final_state, GetPuzzleSize(), curr_state[i]);
 			result = std::pow(std::abs(static_cast<int>((i % GetPuzzleSize()) - goal.first)), 2.0)
 						+ std::pow(std::abs(static_cast<int>((i / GetPuzzleSize()) - goal.second)), 2.0);
-			// std::cout << "EUCLIDE result: " << result << std::endl;
 			if (result != 0.0)
 				dist += std::sqrt(result);
 		}
 	}
-	// std::cout << "EuclideanDist: " << dist << std::endl;
 	return (dist);
 }
